requirements.txt
playwright
pytest-playwright
pytest-html
pytest-bdd
python-dotenv
pytest-xdist
pytest-rerunfailures